fix referrer migration

+++ database/migrations/2022_10_21_113600_add_referrer_id_to_users.php
@@ -13,7 +13,7 @@ return new class() extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table): void {
-            $table->unsignedBigInteger('referrer_id')->nullable();
+            $table->unsignedInteger('referrer_id')->nullable();
             $table->foreign('referrer_id')->references('user_id')->on('users');
         });
     }
